feat: Router now also handle its response
Part of #18

Instead of returning the response object and handle it externally,
Router keeps it, and then handle it in a second method
